requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.1.1
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
entrypoints==0.3
flake8==3.7.8
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
ipykernel==5.1.3
ipython==7.10.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
model-bakery==1.0.2
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
patch==1.16
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.0
psycopg2==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.1
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.3
pyzmq==18.1.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
tblib==1.5.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
Unipath==1.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
